+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.2
 
-- In development directive generators (`export`, `import`, `library`, `part`, `part of`)
+- Implmented directive generators (`export`, `import`, `library`, `part`, `part of`)
 - In development script generators (`library`, `library part`)
 
 ## 0.0.1